bridge server
this include  files which will help to understand the game ui and its
functioning on server side.
Bridge is a playing card game in which max 200 players can play
simultaneously in teams of 2 players as a tournament. All records and
scores will be maintained in server application. which should  be
installed in another android device.